+++ src/components/charts/chartValues/ChartValues.tsx
@@ -1,303 +1,118 @@
-import React, { Component } from 'react'
-import { ViewType } from '../../../config'
-import {
-    getChartValues,
-    updateChartValues,
-    createChartValues,
-    getChartVersionsMin,
-    getChartVersionDetails,
-    deleteChartValues,
-} from '../charts.service'
-import { showError, Progressing, Select, ButtonWithLoader, BreadCrumb, useBreadcrumb } from '../../common'
-import Reload from '../../Reload/Reload'
-import CodeEditor from '../../CodeEditor/CodeEditor'
-import { toast } from 'react-toastify'
-import { ReactComponent as Error } from '../../../assets/icons/ic-warning.svg'
-import { breadCrumbsChartValue, getSavedValuesListURL } from '../charts.helper'
-import { ChartValuesState, ChartValuesProps } from '../charts.types'
-import YAML from 'yaml'
+import React, { useState, useEffect } from 'react'
+import { useParams } from 'react-router'
+import { BreadCrumb, ErrorScreenManager, Progressing, showError, useBreadcrumb } from '../../common'
+import { getChartValuesCategorizedListParsed, getChartVersionDetails, getChartVersionsMin } from '../charts.service'
 import PageHeader from '../../common/header/PageHeader'
+import ChartValuesView from '../../v2/values/chartValuesDiff/ChartValuesView'
+import { ChartInstalledConfig, ChartKind } from '../../v2/values/chartValuesDiff/ChartValuesView.type'
 
-export default class ChartValues extends Component<ChartValuesProps, ChartValuesState> {
-    breadCrumbs = breadCrumbsChartValue(this.props.match.url)
+export default function ChartValues() {
+    const { chartId, chartValueId } = useParams<{ chartId: string; chartValueId: string }>()
+    const [loader, setLoader] = useState(false)
+    const [errorStatusCode, setErrorStatusCode] = useState(0)
+    const [appStoreApplicationName, setAppStoreApplicationName] = useState('')
+    const [valueName, setValueName] = useState('')
+    const [chartInformation, setChartInformation] = useState({
+        appStoreApplicationName: '',
+        deprecated: false,
+        chartName: '',
+        name: '',
+    })
+    const [chartValuesList, setChartValuesList] = useState([])
+    const [availableVersions, setChartVersions] = React.useState([])
+    const [chartValues, setChartValues] = useState({
+        id: 0,
+        kind: null,
+        name: '',
+        chartVersion: '',
+        environmentName: '',
+    })
 
-    constructor(props) {
-        super(props)
-        this.state = {
-            view: ViewType.LOADING,
-            chartVersions: [],
-            chartVersionId: 0,
-            versionData: { id: 0 },
-            name: '',
-            values: '',
-            chartVersion: '',
-            buttonLoader: false,
-            showError: false,
-            appStoreApplicationName: '',
-            isValid: {
-                name: false,
-            },
-        }
-        this.handleChartValuesChange = this.handleChartValuesChange.bind(this)
-        this.handleChartVersionChange = this.handleChartVersionChange.bind(this)
-        this.handleName = this.handleName.bind(this)
-        this.saveChartValues = this.saveChartValues.bind(this)
-        this.redirectToSavedValueList = this.redirectToSavedValueList.bind(this)
-        this.deleteChartValue = this.deleteChartValue.bind(this)
-    }
+    useEffect(() => {
+        getChartDetails()
+    }, [])
 
-    componentDidMount() {
-        getChartVersionsMin(this.props.match.params.chartId)
-            .then((response) => {
-                this.setState({ chartVersions: response.result || [], chartVersionId: response.result[0].id })
-                return response.result[0].id
-            })
-            .then((chartVersionId) => {
-                if (this.props.match.params.chartValueId !== '0') {
-                    getChartValues(this.props.match.params.chartValueId, 'TEMPLATE')
-                        .then((response) => {
-                            this.setState({
-                                name: response.result.name,
-                                chartVersionId: response.result.appStoreVersionId,
-                                values: response.result.values,
-                                chartVersion: response.result.chartVersion,
-                                view: ViewType.FORM,
-                                isValid: {
-                                    name: response.result.name.length > 0,
-                                },
-                            })
-                        })
-                        .catch((error) => {
-                            showError(error)
-                            this.setState({ view: ViewType.ERROR })
-                        })
-                } else {
-                    getChartVersionDetails(chartVersionId)
-                        .then((response) => {
-                            this.setState({
-                                values: response.result.rawValues,
-                                view: ViewType.FORM,
-                                appStoreApplicationName: response.result.appStoreApplicationName,
-                            })
-                        })
-                        .catch((error) => {
-                            showError(error)
-                        })
-                }
-            })
-            .catch((error) => {
-                showError(error)
-                this.setState({ view: ViewType.ERROR })
-            })
-    }
-
-    handleName(event: React.ChangeEvent<HTMLInputElement>): void {
-        let { isValid } = { ...this.state }
-        let name = event.target.value
-        isValid = {
-            name: name.trim().length > 0,
-        }
-        this.setState({ name, isValid })
-    }
-
-    handleChartVersionChange(event: React.ChangeEvent<HTMLFormElement>): void {
-        this.setState({ chartVersionId: event.target.value })
-        getChartVersionDetails(event.target.value)
-            .then((response) => {
-                this.setState({
-                    values: response.result.rawValues,
-                    view: ViewType.FORM,
-                })
-            })
-            .catch((error) => {
-                showError(error)
-            })
-    }
-
-    handleChartValuesChange(valuesYaml: string): void {
-        this.setState({ values: valuesYaml })
-    }
-
-    deleteChartValue(event: React.MouseEvent<HTMLButtonElement>) {
-        deleteChartValues(parseInt(this.props.match.params.chartValueId))
-            .then((response) => {
-                toast.success('Chart Values deleted successfully')
-                this.redirectToSavedValueList(event)
-            })
-            .catch((error) => {
-                showError(error)
-            })
-    }
-
-    redirectToSavedValueList(event: React.MouseEvent<HTMLButtonElement>) {
-        let url = getSavedValuesListURL(this.props.match.params.chartId)
-        this.props.history.push(url)
-    }
-
-    saveChartValues(event: React.MouseEvent<HTMLButtonElement>) {
-        this.setState({ showError: true, buttonLoader: true })
-        let keys = Object.keys(this.state.isValid)
-        let isValid = keys.reduce((isFormValid, key) => {
-            return (isFormValid = isFormValid && this.state.isValid[key])
-        }, true)
-        if (!isValid) {
-            this.setState({ showError: true, buttonLoader: false })
-            return
-        }
+    async function getChartDetails() {
         try {
-            YAML.parse(this.state.values)
+            setLoader(true)
+            const { result: chartVersionMinResult } = await getChartVersionsMin(chartId)
+            setChartVersions(chartVersionMinResult)
+
+            const { result: chartVersionDetail } = await getChartVersionDetails(chartVersionMinResult[0].id)
+            setChartInformation(chartVersionDetail)
+            setAppStoreApplicationName(chartVersionDetail.appStoreApplicationName)
+            const { result } = await getChartValuesCategorizedListParsed(chartId)
+            setChartValuesList(result)
         } catch (error) {
-            toast.error('Invalid yaml')
-            this.setState({ showError: true, buttonLoader: false })
-            return
+            showError(error)
+            setErrorStatusCode(error['code'])
+        } finally {
+            setLoader(false)
         }
-        let promise
-        let requestBody
-        let toastMessage: string
-        if (this.props.match.params.chartValueId !== '0') {
-            requestBody = {
-                id: parseInt(this.props.match.params.chartValueId),
-                name: this.state.name,
-                appStoreVersionId: this.state.chartVersionId,
-                values: this.state.values,
-                chartVersion: this.state.chartVersion,
-            }
-            toastMessage = 'Chart Value Updated'
-            promise = updateChartValues(requestBody)
-        } else {
-            requestBody = {
-                appStoreVersionId: this.state.chartVersionId,
-                name: this.state.name,
-                values: this.state.values,
-            }
-            toastMessage = 'Chart Value Created'
-            promise = createChartValues(requestBody)
-        }
-        promise
-            .then((response) => {
-                toast.success(toastMessage)
-                let { name, chartVersion, values, appStoreVersionId } = response.result
-                this.setState({
-                    name,
-                    chartVersion,
-                    values,
-                    chartVersionId: appStoreVersionId,
-                    buttonLoader: false,
-                    showError: false,
-                    isValid: { name: !!name.length },
-                })
-                this.redirectToSavedValueList(event)
-            })
-            .catch((error) => {
-                showError(error)
-                this.setState({ showError: false, buttonLoader: false })
-            })
     }
 
-    render() {
-        let selectedChartVersion = this.state.chartVersions.find(
-            (chartVersion) => chartVersion.id === this.state.chartVersionId,
+    useEffect(() => {
+        let id, kind
+        if (chartValueId !== '0') {
+            id = parseInt(chartValueId)
+            kind = ChartKind.TEMPLATE
+        } else {
+            id = availableVersions[0]?.id
+            kind = ChartKind.DEFAULT
+        }
+        if (id) {
+            const chartValues = chartValuesList.find((chrtValue) => {
+                if (chrtValue.kind === kind && chrtValue.id === id) return chrtValue
+            })
+            if (chartValues) {
+                setChartValues(chartValues)
+                if (chartValueId !== '0') {
+                    setValueName(chartValues.name)
+                }
+            }
+        }
+    }, [availableVersions, chartValuesList])
+
+    if (loader) {
+        return <Progressing pageLoader />
+    }
+    if (errorStatusCode > 0) {
+        return (
+            <div className="error-screen-wrapper flex column h-100">
+                <ErrorScreenManager
+                    code={errorStatusCode}
+                    subtitle="Information on this page is available only to superadmin users."
+                />
+            </div>
         )
-        if (this.state.view === ViewType.LOADING) {
-            return <Progressing pageLoader />
-        } else if (this.state.view === ViewType.ERROR) {
-            return <Reload />
-        } else {
-            return (
-                <div className="chart-values">
-                    <Header appStoreApplicationName={this.state.appStoreApplicationName} name={this.state.name} />
-                    <form className="chart-values__container">
-                        <label className="form__row">
-                            <span className="form__label">Name*</span>
-                            <input
-                                tabIndex={1}
-                                placeholder="Name"
-                                className="form__input"
-                                disabled={this.props.match.params.chartValueId !== '0'}
-                                value={this.state.name}
-                                onChange={this.handleName}
-                            />
-                            {this.state.showError && !this.state.isValid.name ? (
-                                <span className="form__error">
-                                    <Error className="form__icon form__icon--error" />
-                                    This is a required field
-                                </span>
-                            ) : null}
-                        </label>
-                        <div className="form__row">
-                            <span className="form__label">Chart version*</span>
-                            <Select
-                                tabIndex={2}
-                                value={this.state.chartVersionId}
-                                onChange={this.handleChartVersionChange}
-                            >
-                                <Select.Button rootClassName="select-button--default">
-                                    {selectedChartVersion ? `${selectedChartVersion.version}` : 'Select Version'}
-                                </Select.Button>
-                                {this.state.chartVersions.map((chartVersion, index) => {
-                                    return (
-                                        <Select.Option key={`${chartVersion.version}`} value={chartVersion.id}>
-                                            {chartVersion.version}
-                                        </Select.Option>
-                                    )
-                                })}
-                            </Select>
-                        </div>
-                        <div className="mb-12 code-editor-container">
-                            <CodeEditor
-                                value={this.state.values}
-                                noParsing
-                                mode="yaml"
-                                onChange={this.handleChartValuesChange}
-                            >
-                                <CodeEditor.Header>
-                                    <span className="bold">values.yaml</span>
-                                </CodeEditor.Header>
-                            </CodeEditor>
-                        </div>
-
-                        <div className="flex right">
-                            {this.props.match.params.chartValueId !== '0' ? (
-                                <button type="button" className="cta delete mr-16" onClick={this.deleteChartValue}>
-                                    Delete
-                                </button>
-                            ) : (
-                                <button
-                                    type="button"
-                                    className="cta cancel mr-16"
-                                    onClick={this.redirectToSavedValueList}
-                                >
-                                    Cancel
-                                </button>
-                            )}
-                            <ButtonWithLoader
-                                isLoading={this.state.buttonLoader}
-                                rootClassName="cta"
-                                loaderColor="white"
-                                onClick={this.saveChartValues}
-                            >
-                                Save Template
-                            </ButtonWithLoader>
-                        </div>
-                    </form>
-                </div>
-            )
-        }
     }
+    return (
+        <>
+            <Header appStoreApplicationName={appStoreApplicationName} name={valueName} />
+            <ChartValuesView
+                isCreateValueView={true}
+                installedConfigFromParent={chartInformation as ChartInstalledConfig}
+                chartValuesListFromParent={chartValuesList}
+                chartVersionsDataFromParent={availableVersions}
+                chartValuesFromParent={chartValues}
+                selectedVersionFromParent={availableVersions[0]?.id}
+            />
+        </>
+    )
 }
 
 function Header({ appStoreApplicationName, name }) {
     const { breadcrumbs } = useBreadcrumb(
         {
             alias: {
-                ':chartId': { component: appStoreApplicationName || null, linked: true },
+                ':chartId': appStoreApplicationName || null,
                 chart: null,
                 ':chartValueId': { component: name || 'New value', linked: false },
                 'chart-store': null,
-                'saved-value': 'Saved value',
+                'saved-values': 'Saved value',
             },
         },
-        [],
+        [appStoreApplicationName, name],
     )
 
     const renderChartValueBreadcrumbs = () => {
